+++ TwitchChannelPointsMiner/__init__.py
@@ -14,6 +14,7 @@ from .status_dashboard_patch import apply_patch as _apply_status_dashboard_patch
 from .status_dashboard_webhook_patch import apply_patch as _apply_status_dashboard_webhook_patch
 from .status_dashboard_enhancements_patch import apply_patch as _apply_status_dashboard_enhancements_patch
 from .status_dashboard_clock_patch import apply_patch as _apply_status_dashboard_clock_patch
+from .status_dashboard_error_safety_patch import apply_patch as _apply_status_dashboard_error_safety_patch
 from .watch_notifications_patch import apply_patch as _apply_watch_notifications_patch
 from .watch_streak_persistence_patch import apply_patch as _apply_watch_streak_persistence_patch
 
@@ -29,6 +30,7 @@ _apply_status_dashboard_patch()
 _apply_status_dashboard_webhook_patch()
 _apply_status_dashboard_enhancements_patch()
 _apply_status_dashboard_clock_patch()
+_apply_status_dashboard_error_safety_patch()
 del _apply_channel_points_context_patch
 del _apply_discord_format_patch
 del _apply_drop_game_main_list_preference_patch
@@ -37,6 +39,7 @@ del _apply_fork_branding_patch
 del _apply_playback_access_token_patch
 del _apply_status_dashboard_clock_patch
 del _apply_status_dashboard_enhancements_patch
+del _apply_status_dashboard_error_safety_patch
 del _apply_status_dashboard_patch
 del _apply_status_dashboard_webhook_patch
 del _apply_watch_notifications_patch